requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
distlib==0.3.7
Django==4.2.9
django-bootstrap-v5==1.0.11
django-bootstrap5==23.3
djangorestframework==3.14.0
filelock==3.12.3
platformdirs==3.10.0
psycopg==3.1.18
psycopg2==2.9.2
psycopg2-binary==2.9.9
pytz==2023.3.post1
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
virtualenv==20.24.5
